milestone4/requirements.txt
shiny==0.9.0
shinyswatch==0.6.0
seaborn==0.12.2
matplotlib==3.7.1
pandas
geopy
faicons
shinywidgets
ipyleaflet
asyncio
numpy
transformers
Pillow
torch